# Improve performance of exec-hash