--- a/docs/docs/install/config/index.md
+++ b/docs/docs/install/config/index.md
@@ -134,7 +134,7 @@ output:
         stack-addresses: true
         exec-env: false
         relative-time: true
-        exec-hash: false
+        exec-hash: dev-inode
         parse-arguments: true
         sort-events: false