--- a/pkg/config/config.go
+++ b/pkg/config/config.go
@@ -153,11 +153,35 @@ type CapabilitiesConfig struct {
 // Output
 //
 
+type CalcHashesOption int
+
+const (
+	CalcHashesNone CalcHashesOption = iota
+	CalcHashesPathname
+	CalcHashesDevInode
+	CalcHashesDigestInode
+)
+
+func (c CalcHashesOption) String() string {
+	switch c {
+	case CalcHashesNone:
+		return "none"
+	case CalcHashesPathname:
+		return "pathname"
+	case CalcHashesDevInode:
+		return "dev-inode"
+	case CalcHashesDigestInode:
+		return "digest-inode"
+	default:
+		return "unknown"
+	}
+}
+
 type OutputConfig struct {
 	StackAddresses bool
 	ExecEnv        bool
 	RelativeTime   bool
-	CalcHashes     bool
+	CalcHashes     CalcHashesOption
 
 	ParseArguments    bool
 	ParseArgumentsFDs bool