--- a/pkg/ebpf/tracee.go
+++ b/pkg/ebpf/tracee.go
@@ -1569,6 +1569,7 @@ func (t *Tracee) getFileHash(filename string, ctime int64, mountNs int, containe
 		if err == nil {
 			hashInfoObj = fileExecInfo{ctime, hash}
 			t.fileHashes.Add(fileId, hashInfoObj)
+			fileHash = hash
 		}
 	}
 	return fileHash, nil