Close: #3734 
Close: #3745

### 1. Explain what the PR does

f24813f12 **feat(ebpf): expand exec-hash output option**
43f46089b **feat(ebpf): optimize computeFileHash**
e90ad8576 **fix(ebpf): fix getFileHash()**
7fcc7341b **fix(bucketscache): GetBucket was not thread safe**
4d04a15d4 **chore(bucketscache): add benchmark**


f24813f12 **feat(ebpf): expand exec-hash output option**

```
The user can now choose between the following options:

- none (default)
- pathname
- dev-inode
- digest-inode

'pathname' option recalculates the file hash if the inode's creation
time (ctime) differs, which can occur in different namespaces even for
identical pathnames.

'dev-inode' option generally offers better performance compared to the
pathname option, as it bypasses the need for recalculation by
associating the creation time (ctime) with the device (dev) and inode
pair.

'digest-inode' option is the most efficient, as it keys the hash to a
pair consisting of the container image digest and inode. This approach,
however, necessitates container enrichment.
```


43f46089b **feat(ebpf): optimize computeFileHash**

```
Improve the computeFileHash function, by using minio/sha256-simd.

Benchmarking indicates the current implementation is nearly three times
faster than the old version (21,311 ns/op vs. 61,857 ns/op) while
maintaining similar memory efficiency (33,040 B/op vs. 33,056 B/op) and
the same number of memory allocations per operation (5 allocs/op).

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/ebpf
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
BenchmarkComputeFileHashOld-20             19257             61857 ns/op           33056 B/op          5 allocs/op
BenchmarkComputeFileHashCurrent-20         56769             21311 ns/op           33040 B/op          5 allocs/op
```


e90ad8576 **fix(ebpf): fix getFileHash()**

```
First call to getFileHash() for a given filename was returning an
empty string.

Context: #3745
```


7fcc7341b **fix(bucketscache): GetBucket was not thread safe**

```
GetBucket was not thread safe since it was returning a reference to a
bucket instead of its copy.
```


4d04a15d4 **chore(bucketscache): add benchmark**

```
It demonstrate that the current algorithm of addBucketItem with multiple
lock/unlock) is optimal compared to a single lock/unlock.

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/bucketscache
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
BenchmarkAddBucketItemCurrent-20                     464           2586038 ns/op           2118 B/op           5 allocs/op
BenchmarkAddBucketItemWithOneLock-20                 122           9461730 ns/op            424 B/op           1 allocs/op
```

### 2. Explain how to test it

### 3. Other comments